Add PostCSS import plugin as default
https://github.com/postcss/postcss-import

It felt natural to me to immediately use the @import directive and after fighting a bit with it I relised the plugin wasn't installed so it seemed like it would be a sensible default to have.

Adding the official postcss-import plugin enables the usage of PostCSS's @import directive out of the box so when adding a CSS library (like open-props) or when adding custom stylesheets it works seamlessly

Note that this is **not** the CSS native @import.